refactor(dispatch): decompose into per-integration modules
Extract each integration (github, gitlab, linear, things) into its own
module implementing a shared Integration interface. Sources, context,
and resolve become thin dispatchers over the integration registry.
