+++ packages/dispatch/run.ts
@@ -1,0 +1,10 @@
+export async function run(cmd: string[]): Promise<string> {
+  const proc = Bun.spawn(cmd, { stdout: "pipe", stderr: "pipe" });
+  const exitCode = await proc.exited;
+  if (exitCode !== 0) {
+    const stderr = await new Response(proc.stderr).text();
+    throw new Error(`Command failed: ${cmd.join(" ")}\n${stderr}`);
+  }
+  const stdout = await new Response(proc.stdout).text();
+  return stdout.trim();
+}
